pyproject.toml
[project]
name = "inference-sawit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "lap>=0.5.12",
    "opencv-python>=4.12.0.88",
    "pytz>=2025.2",
    "shapely>=2.1.2",
    "ultralytics>=8.3.221",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "psycopg2-binary>=2.9.10",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "openpyxl>=3.1.5",
    "flask-migrate>=4.1.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128"},
]
torchvision = [
  { index = "pytorch-cu128"},
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
]
